# src/docs/swagger.json
{
    "swagger": "2.0",
    "basePath": "/api",
    "host" : "localhost:3000",
...
    "consumes" : [
        "application/json"
    ],
    "schemes" : [
        "http"
    ],
    "produces" : [
        "application/json"
    ],
    "info": {
        "version" : "1.0.0",
        "description" : "Node JS Swagger API for Oracle HR database",
        "title" : "Node API",
        "contact": {
            "name": "Daniel Bennin",
            "email": "danyelamps.db@gmail.com"
          }
    },
    "paths": {
...
        "/regions/updateRegion/{regionId}" : {
            "put" : {
                "tags": [
                    "Regions"
                  ],
                "description" : "Update a Region from Oracle HR Database",
                "summary" : "Update Region",
                "parameters" : [
                    {
                        "name" : "regionId",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "description" : "Region Id"
                    },
                    {
                        "name" : "regionName",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "description" : "Region Name",
                        "schema" : {
                            "$ref" : "#definitions/Region"
                        }
                    }
                ],
                "responses" : {
                    "400" : {
                        "description" : "Update Failed."
                    },
                    "200" : {
                        "description" : "Update Successful"
                    }
                }
            }
        },
        "/regions/removeRegion/{regionId}" : {
            "delete" : {
                "tags": [
                    "Regions"
                  ],
                "description" : "Remove a Region from Oracle HR Database",
                "summary" : "Delete Region",
                "parameters" : [
                    {
                        "name" : "regionId",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "description" : "Region Id"
                    }
                ],
                "responses" : {
                    "400" : {
                        "description" : "Delete Failed."
                    },
                    "404" : {
                        "description" : "Delete Successful"
                    }
                }
            }
        },
        "/regions/addRegion" : {
            "post" : {
                "tags": [
                    "Regions"
                  ],
                "description" : "Add a new Region to Oracle HR Database",
                "summary" : "Add new Region",
                "parameters" : [
                    {
                        "name" : "regionName",
                        "in": "body",
                        "required": true,
                        "type": "string",
                        "description" : "Department Id",
                        "schema" : {
                            "$ref" : "#definitions/Region"
                        }
                    }
                ],
                "responses" : {
                    "201" : {
                        "description" : "Created"
                    },
                    "400" : {
                        "description" : "Bad Request"
                    },
                    "500" : {
                        "description" : "Insert Unsuccessful"
                    }
                }
            }
        }
    }
}